Autoupdate of 2 submodules
*art updated to cda3d47149b41d85c38d1f3ffbc312a930b147b2
Update Art Service to be a LocalManager

This CL re-names the ArtService class to conform to the Frameworks
naming convention for local managers, which are SystemServer only
"service" singletons which are accessible through the LocalManagers
registry.  The build system is updated for these changes, as well as
adding appropriate API tracking build targets.

Test: build, boot, and inspect log
Bug: 177273468
Merged-In: Iec3d2530a3a3a2824009b50f13e297da8fde55c5
Change-Id: Iec3d2530a3a3a2824009b50f13e297da8fde55c5

*build updated to 8c614d25159cf9db2956b01080d487300985f9ce
Merge "Add the service-art JAR to the System Server classpath."