Revert^2 "[metrics] Report CompilationReason and CompilerFilter"
Since we do not know the real primary OAT file when the runtime starts
up, we have to wait until NotifyStartupCompleted to be sure we have
the right one. To accomodate this, we delay starting the metrics
session until the first report and then send updated session
information in NotifyStartupCompleted.

Note that we could delay the whole metrics reporting initialization to
NotifyStartupCompleted, but this doesn't get called in all
circumstances, which would mean metrics reporting is disabled in cases
where we want it.

This reverts commit 8b07c653a7e45886ee92d8a12aa439e98e84befa.

Reason for revert: Fixing crash

The bug was get GetPrimaryOatFile could return null in some cases, but
we were not handling this correctly.

Bug: 178239096
Test: manual
Change-Id: I468837ee8f026edd82c47bd216ddb5b2d9c2ece3