vts_treble_vintf_*_test: static deps
A dependency to the APEX interface was added recently using
libapexservice-deps, but this has a shared lib dependency. Normally, I'd
create something like 'libapexservice-deps-static'. However, the
dependencies of an AIDL lib are now and will remain libbinder+libutils
(additional dependencies must be handled by aidl_interface.go). So, I am
using an explicit dependency on the AIDL interface here, as is standard
practice.

Fixes: 184616089
Test: inspect `readelf -d .../vts_treble_vintf_vendor_test`
Change-Id: Iafac89c2419a2519c3d1fc6bbb6b20852332c6c6