Add support for deterministic DSA signing
This commit adds support for deterministic DSA signing (RFC6979)
using the Java signature algorithm SHA256withDetDSA. Note that this
signature algorithm may not be supported by the default Security
Providers and may require adding an additional Provider. The
deterministic signature is still a valid signature and can be
verified using the same process to verify a standard DSA signature.

Bug: 169266066
Test: gradlew test
Change-Id: I80bd17302814411c0feaea8a7e6a08e80476d7ef