Add a --test-exec flag to bisect_profile.py
Often one has a small script that can tell whether a bisection has the
target behavior or not. Add a --test-exec flag to allow
bisect_profile.py to drive itself using a non-zero exit to indicate
that the interesting behavior is present and a 0 exit to indicate it
is not.

Test: ./art/tools/bisect_profile.py --apk ~/GoogleExtServices.apk --test-exec './art/tools/compile-jar.py --profile-file bad.prof --arch arm64 ~/GoogleExtServices.apk --force-allow-oj-inlines -j1' --output-source bad.txt bad.prof
Change-Id: I2fbaf15da8177953ad0051274ee974e6c0e9f503