VtsSecurityAvbTest: Adding GKI 2.0 test
In GKI 2.0, a device should use a generic boot image as the
shipping image. To do GKI 2.0 verificaiton, a boot_signature
is added to boot image v4, where the boot_signature can be
used to check the intergrity of the kernel and the ramdisk.

This CL adds the verification test for GKI 2.0, if the kernel
version is >= 5.10. On the other hand, if the kernel version is 5.4,
it performs GKI 1.0 verification by using the boot vbmeta from
the end of the boot image.

Bug: 177862434
Test: atest vts_security_avb_test
Change-Id: I3cc25f555c94fcaeff7caff284406b7773dece66