Autoupdate of 2 submodules
*build/blueprint updated to 78df853e07f5fe6bed70daf72357900532d99033
Make it possible to call the primary builder more than once.

This is accomplished by plumbing PrimaryBuilderInvocation in from
bootstrap.Args.

The confusing removal of "primaryBuilderExtraFlags" from documentation
generation is because it used to (erroneously) refer to the func defined
in command.go and its removal apparently doesn't break anything. I
suspect that this code path is dead within Android, but we'll need to
wait until May to remove it.

Test: Presubmits.
Change-Id: I8763579b82bd79d5ee0f233edfd48327282e401c

*build/soong updated to 3d76d5d6b8aae617572d4a2d6b2d8366c035acd9
Merge "Allow running bp2build as part of a regular build."