Cache Symbols::GetGlobal
The ART tests currently time out in this method after 24 hours.

There is other unknown failure in unwinding of some frequent frame.
This causes the unwinder to try JIT as well, however no JIT is used
at all in those tests, which means the JIT is being re-initialized
on excessive number of unwinds. In particular, taking a lot of time
in the linear scan over symbols to find the magic global variable.

There might be some better way to handle the JIT initialization,
but the caching of the symbol lookup is good optimization anyway.

Test: libunwindstack_unit_test
Change-Id: I6b3d9a03b80da4d71a5e6bc61c551fe94a277164