Autoupdate of 2 submodules
*hardware/interfaces updated to 4b84c912a006f6614dc82da69c2c19850db2e05e
Merge "Porting IRPC functionality."
*system/keymaster updated to 57c187a4d14d6b23238aa46f55bff4d62d2e0125
Migrating remote provisioning service functionality

The initial implementation of the IRemotelyProvisionedComponent HAL had
everything being done in the userspace portion of the process, including
the sensitive crypto. This change ports that functionality to
system/keymaster. This includes changes to the following components to
work in this module to AndroidKeymaster:

-Additions to keymaster_messages to support the two new methods:
 GenerateCsr and GenerateRkpKey.

-The creation of a new RemoteProvisioningContext class to abstract away
 the implementation specific details from AndroidKeymaster.

-A handful of added utility functions/files.

Test: atest VtsHalRemotelyProvisionedComponentTargetTest
Change-Id: Ic901cea79e075a170635cb3697c065fbbd020417