Rename StatsD/bin to Statsd/statsd
statsd is a more intuitive name for this, as it is the files for the
stats daemon. It also matches the previous structure of
frameworks/base/cmds/statsd.

There may be some confusion between StatsD (in packages/modules/StatsD)
and the inner statsd. The difference is that StatsD is the name of the
apex, containing statsd but also the supporting libraries and jars. The
lowercase statsd is for statsd the daemon.

Test: m
Bug: 182597781
Change-Id: Ibb065d64be3775c699583215f7f376a1a5492ac1