Add installclean step to buildbots
There can be leftover files from previous incremental build.
This is usually harmless, since we rebuild everything we use.

However, in some cases, presence of stale files can be confusing.
For example, if test does not need dex2oat, but will use one if
present (which might be stale now), or if we there are libraries
in the build which have been deleted in previous CLs.

Bug: 133142426
Test: ./art/test/testrunner/run_build_test_target.py art-jit
Change-Id: Iede85b568a7c56de4dc80f488fe09772ee767a37