libunwindstack: try stripping PAC from pc
This patch introduces a general method for "fallback" pc values, used
in case the starting pc points to unmapped memory. This can be useful
in cases where the crash was caused by the program jumping to a bad
address.

For Arm64, this fallback value is the pc with the PAC extension bits
cleared. This is useful in situations where the program jumped to
addresses that still had something in the PAC bits. This can happen
when a signed pointer was not authenticated, or the authentication
failed. Without the fallback, the backtrace would only see the address
where the program crashed. Because of the PAC bits, it cannot point to
any mapped memory locations, which immediately stops the
backtrace. With the fallback, it is able to produce a useful
backtrace.

Test: libunwindstack_test

Change-Id: I0cb204ab2bee8b78978d348c496e31f63228a926