Autoupdate of 2 submodules
*external/crosvm updated to b36408f79c53537b851351b2022c2818388c4903
Fix build files

Bug: 185155959
Test: locally
Change-Id: I53ce0b62c3e18e2edfe43aa36d13bc29e039f5ac

*external/rust/crates/libc updated to 4c174715e21655b2c1bcacb15a73931523d421bf
Add functions and constants crosvm depends on to the android targets

Rust's Libc doesn't expose some functions and constants for android
even though those are present in bionic and available to native code.
This change temporarily adds the ones that crosvm depends on until the
problem is fixed upstream.

Bug: 185155959
Test: locally
Change-Id: Ib955ba8ced45580561e50afc5d77a2e9b7170168