Autoupdate of 4 submodules
*external/crosvm updated to b36408f79c53537b851351b2022c2818388c4903
Fix build files

Bug: 185155959
Test: locally
Change-Id: I53ce0b62c3e18e2edfe43aa36d13bc29e039f5ac

*external/rust/crates/libc updated to 4c174715e21655b2c1bcacb15a73931523d421bf
Add functions and constants crosvm depends on to the android targets

Rust's Libc doesn't expose some functions and constants for android
even though those are present in bionic and available to native code.
This change temporarily adds the ones that crosvm depends on until the
problem is fixed upstream.

Bug: 185155959
Test: locally
Change-Id: Ib955ba8ced45580561e50afc5d77a2e9b7170168

*external/rust/crates/smallvec updated to 784e24eb3f0c100bb3f2da234330b8d93b27b722
Configure apex so crosvm can depend on this crate

Bug: 185155959
Test: locally
Change-Id: I2ba7694d4d511e64efc2971dc5549d540398c801

*external/rust/crates/vmm_vhost updated to 200df3137aaddb6fc65e643ed394eeb8097c50fe
Re-Add Android.bp file

This time in a topic with the crosvm changes it depends on

This reverts commit 66cbf7119a2e3023664171adbd528ac94243e2c5.

Reason for revert: Merging along the changes it depends on

Change-Id: Idafa87fcc683efece89fdd807117b12e111bb1ef