Do not de-dupe feature removed requests when the server is unavailable
When UNAVAILABLE_REASON_SERVER_UNAVAILABLE is returned in imsFeatureRemoved,
the connection to the telephony process itself is not available and
connect() will need to be retried. In this case, we do not want to
de-dupe the same state updates to connectionUnavailable, as this
will require the caller to retry using connect() again.

Fixes: 186288282
Test: atest ImsCommonTests
Change-Id: I5a683543ebc9e3ea2add6e311853cae827a448a6