Throw an error on all overlay application errors
libufdt depends on the presence of the __overlay__ subnode to identify a
node in an overlay DT as a fragment node. Since base DT target only
needs to resolved for valid fragment nodes, return early for
non-fragment nodes. This allows us to return an error code for any other
error.

Test: sh tests/run_tests.sh
Bug: 186482257
Change-Id: I05ec377e3773dc4a6396278f246da994c9004865