power_profile.xml: Fix typo "volatage"
In the before-times of June 2019, an intrepid AOSP reviewer named
"fbechet45@gmail.com" reported:

> typo error:
> line 311 : "volatage" should be written "voltage"

After a short series of hops, this bug landed in my todo list.

My first thought was, "Easy!  I'll fix this on my internal git,
Lets go, in-and-out, 20 minute adventure."  But, that would deny the
original reporter of seeing the solution for some large number of
months.

So, I decide "I will fix this in AOSP.  But, oh no, Coral has the
same copy-paste error!  Let me just wait until October 2019 when
Coral goes public, then I'll fix those two at the same time."

(Little did I know this would continue to get copied and pasted, and
now there are 20+ instances of this on our internal repository.)

Fast forward to 2020: The pandemic gripped our nation, and I was
busy baking bread and posting pictures of it on Facebook.  I have
some great recipes I can share.  Let me know if you need some of my
artisinal sourdough starter.

Now, in 2021, after multiple automated-and-human reminders, this
has become my #1 priority issue to solve.  I still have the "can't
fix all the errors" on AOSP problem, but if I don't fix crosshatch
now, it will be too late to land it into the 2021 dessert release,
which may be the last supported release for crosshatch.  Oh no!

I am determined not to mark this bug as "Won't Fix / Obsolete"
and yet not so determined to act on it quickly.  So, I will fix
crosshatch internally, and then fix the rest of the projects on our
internal master branch.

I hope this patch finds you well and remember to take care of
yourself in these tough times.

With regards,
Chris

Bug: 132925372
Bug: 135906639
Bug: 77804134
Change-Id: I2d088bcbdfbc758eed5054f0413ab02aa5e57e4c