Allow -unstable in BOARD_KERNEL_MODULE_INTERFACE_VERSIONS.
Add packages that corresponds to the -unstable generation
for KMI versions that has not been frozen yet.

Update build time checks and tests to parse the -unstable
generation correctly. When -unstable is specified in a
KMI version in BOARD_KERNEL_MODULE_INTERFACE_VERSIONS or
APEX package name, it can match any generation.

Test: GkiInstallTest
Bug: 187432172

Change-Id: Id20e6a643bb9a48af378f53026bd95b1b9d7e604