Support fetching from an upstream main git branch
The external updater currently assumes the name of the upstream git
branch to fetch is always "master".  This is no longer true, as some
repositories use "main" instead.  This commit lets the updater try to
use both of those branches.

Note that there are still a few instances of "master" in the updater,
but I believe those correspond to AOSP and GitHub specifically and so
shouldn't need to be changed yet.  If/when they do, we should be able
to change them easily with these changes.

As part of this change we also fix a potential bug in a
previously-unused function so that it does not strip part of the
branch name from a string of the form "<remote name>/<branch name>".

Test: Run the updater with upstream master and upstream main.
Change-Id: I0cf83dbf1765ff26b602176fa1c1ff95224e1a7f