Export dexdump, oatdump, and veridex as prebuilt host tools.
dexdump is used with the check_boot_jars script when building boot
jars.

oatdump is useful for debugging. In particular having it as a prebuilt
makes it easier for external parties to find the right oatdump binary -
the format isn't stable so it's important to use the version
corresponding to the APEX package.

veridex is used for appcompat checks on system image apps.

Test: art/build/build-art-module.sh --skip-apex
      packages/modules/ArtPrebuilt/update-art-module-prebuilts.py \
        --local-dist out/dist --skip-apex
      ldd prebuilts/module_sdk/art/current/host-exports/x86_64/bin/dexdump
      ldd prebuilts/module_sdk/art/current/host-exports/x86_64/bin/oatdump
      ldd prebuilts/module_sdk/art/current/host-exports/x86_64/bin/veridex
      m SOONG_CONFIG_art_module_source_build=false droid
Bug: 172480615
Change-Id: Ice0583089777f40b9b12c1a602ebc56fc22f9335