Clean up ota_apex when we reserve 0 bytes for CAPEX
If OTA contains CAPEX, then depending on situation we reserve some space
on device for decompression. Later, as part of post-install hook of OTA,
we decompress the CAPEX inside OTA from chroot.

If the user decides to cancel the OTA, they call the API on apexd for
reserving space with a 0. That deletes the space we have reserved.

If user is canceling the OTA and cleaning up the reserved space, we
should also clean up the ota_apex we created in post-install hook. Since
OTA is cancelled, there is no use for these ota_apex anymore.

Bug: 172911822
Test: atest ApexTestCases
Change-Id: Iba26bafa54a98a96513dad093899c3b721c99f85