Do not double check trace config targetting.
heapprofd now includes the target tracing session ID when signalling the
ART plugin to initiate a dump. This makes the code we used to need to
figure out whether a tracing session targets the current process
redundant.

The old code also had a bug, because it checked the requested PID
against getpid(). As this was done after the call to fork(), that broke
targetting by PID.

Test: run concurrent perfetto_hprof sessions and verify the results are
      as expected.

Bug: 188523717
Merged-In: I2f68380f88f10709248fc2cc268bb24db359e6c5
Change-Id: I2f68380f88f10709248fc2cc268bb24db359e6c5