Autoupdate of 2 submodules
*build/bazel updated to c2a46c4dc86bc3de55724930a40aa6c8d4a04bb0
Merge "Add rtti to library features based on rtti bool"
*prebuilts/clang/host/linux-x86 updated to c67bd442408819743a1f02ff09e42017834782ed
Add rtti feature to cc toolchain

This adds `-frtti` if "rtti" is added to a library's feature list, and
`-fno-rtti` otherwise. This has parity with Soong's rtti property.

Test: Manually verified command line contains appropriate value with one
rtti case and one non-rtti case
Test: CI scripts
Test: Verified this unblocked several modules from mixed build denylist,
to be removed in a later CL.

Change-Id: Idd7b983c0e154434f71b46f107cd493c5af6070e