testrunner.py: Add --build-only and --skip-build options.
The build stage is more expensive than the test stage
and we do it redundantly repeatedly for all variants.

Make it possible to separate the build and test stages,
which will allow us to do the build just once on buildbots.

The proper solution is to move all the build work to soong,
and we will be able to remove this code again at that point.

Bug: 188631922
Test: testrunner.py --host --build-only && \
      testrunner.py --host --skip-build
Change-Id: I542feac03acc25a853dbf7f1a2e5587a2c5d6d7a