Replace @link with @code for public alternatives
Hidden API modularization causes the existing mechanism to verify the
validity of publicAlternatives specified in an UnsupportedAppUsage
annotation to break when the alternative API is provided by a separate
bootclasspath_fragment. That is because it can see the APIs provided
by the current bootclasspath_fragment.

This circumvents that by replacing the @link doctag with an @code
doctag which does not perform the checking. Bug 187693462 covers the
work to fix this issue and once that has been done then this can be
reverted.

Bug: 187693462
Bug: 179354495
Test: - modify external/conscrypt/apex to perform hidden API processing
        in the com.android.conscrypt-bootclasspath-fragment.
      m out/soong/hiddenapi/hiddenapi-flags.csv
      - Before this change the build fails because it cannot find the
        alternative APIs which are provided by art-bootclasspath-fragment
        After this change the build passes.
Change-Id: Ieb2dc229d3e56d50d63568de56c9b66e1c799ee7