Autoupdate of 3 submodules
*art updated to 6d8d68efbe2101e2d5fdf078556cbc1f286c37c7
Support modularization in "hiddenapi list" tool

Previously, the "hiddenapi list" tool assumed that it was generating
flags for the whole bootclasspath and so did not need to differentiate
between classes that are provided by the bootclasspath fragment and
classes that are simply used by it. That meant that it would output
flags for this bootclasspath fragment plus all the other parts of the
bootclasspath on which this depends.

e.g. When generating the flags for i18n it would also output the flags
for art too.

This change adds the --dependency-stub-dex option to specify those dex
files that contain the API stubs provided by other bootclasspath
fragments and used by this one. A file specified in the
--dependency-stub-dex option is treated in a similar way to the files
added to the --boot-dex but its members are ignored when generating the
output file.

Test: - enable hiddenapi processing in migrate packages/modules/RuntimeI18N
      m out/soong/hiddenapi/hiddenapi-flags.csv
      - the previous command will verify that the I18N's generated
        all-stubs.flags matches the subset of monolithic flags that overlap.
      - check the I18N's generated all-stubs.flags to make sure it is
        complete.
Change-Id: I7307500eff1141d161625ec69696802ff3ab82a9

*build/soong updated to f1b358cb5764039ff7502e9f73ad0abfdc91c7aa
Support hidden API processing for fragments with dependencies

Previously, a bootclasspath_fragment that depended on classes provided
by another bootclasspath_fragment did not support hidden API processing
as it would not supply information about those dependencies.

This change adds support for that as follows. Each fragment:
1. Exports the transitive sets of stub dex jars for each of the public,
   system, test and core_platform APIs (where relevant).
2. Adds dependencies onto its dependent fragments.
3. Retrieves the API stubs dex jars from its dependent fragments and
   passes them to the "hiddenapi list" tool which will use them to
   resolve dependencies but will not output them to the generated
   flags.

Once the flags are generated the existing encoding functionality
encodes the flags into the dex files of the bootclasspath_fragment's
content modules which are then packaged into the apex.

Bug: 179354495
Test: m com.android.sdkext
      - verify that this does not change the contents of the apex files
Change-Id: I3e82a6dbb437f1e417e5d7e25aeb212e378603d0

*packages/modules/SdkExtensions updated to 7cc1a94355241e8457d3062c0981b938dee9d9c7
Merge "Enable hidden API processing in com.android.sdkext-bootclasspath-fragment"