Switch test_com.android.sdkext to use java_sdk_library
Previously, due to limitations in the build's ability to perform hidden
API encoding on its contents the test_com.android.sdkext apex used a
dex_import module that was built by the testing/update_prebuilt.sh.
That limitation no longer exists and so this change switches it to use
java_sdk_library.

Bug: 179354495
Test: m com.android.sdkext test_com.android.sdkext
      - dexdump the contained framework-extensions.jar of both
      - compare the output and make sure that the hiddenapi flags are
        correct.
Change-Id: I76ba8f334a49214d3039ffb7702156741d7a4fe7