Autoupdate of 2 submodules
*system/apex updated to 8cef66e2b9f3c2cbd6b8654fa8d12c16ee8687aa
Basic implementation of installAndActivatePackage binder call

To make it easier to review this CL just implements the basic flow.
Following will be implemented in the follow-up CL(s):

* Error handling and clean up
* e2e tests

Since some binaries that are using an APEX that is being updated can run
in a separate mount namespace, we can't just destroy dm-verity and loop
devices that previously active version of APEX is mounted on top of.
Instead, we are making use of the DM_DEFERRED_REMOVE and
LO_FLAGS_AUTOCLEAR flags which will automatically tear down dm and loop
block devices respectively after last close.

Due to the same reasons, supporting a samegrade becomes a little bit
tricky: we need to pick up a name for the device that doesn't collide
with one that is backing a previous APEX that is still accessed by a
process in it's own mount namespace (same applies to the final location
of APEX on /data/apex/active).

To handle this a concept of "minor version" is introduced: name of a
dm-verity device now follows ${module_name}@{version_code}[_${minor}]
pattern. A minor version of the new dm-verity device is calculated as
max(minor versions of other dm-verity devices with same
${module_name}) + 1.

Test: atest ApexTestCases
Test: adb install --force-non-staged rebootless_apex_v2.apex
Bug: 187864524
Bug: 188713178
Change-Id: I346ecf1360be87320c718922130261f6f244f1ac

*system/core updated to 45acb369eec9b632e2a3ea44e9600f6014f9bf7b
Merge "libdm: Add DeleteDeviceDeferred API"