Fill the audio vbuffer to prevent underrun
Before the patch, the audio vbuffer for output is not fully used because
of the "simulated sleep" in the out_write function. In this patch, we
keep the audio vbuffer full to prevent underrun.

This patch can also help us remove the "buffer size * 2" hack
introduced in I6feda36b4e961539555e082e2e6b2d59c73d8f7f

Bug: 184577782

Test: Build; flash; play sounds; quickly switch musics
Change-Id: I0025b0aa2730ed98089d4bac4926630512ac7662