Autoupdate of 4 submodules
*external/boringssl updated to 79b5cad203d328cec256e0a6cf1319f215d8d41f
[trusty] Add boringssl to library build system

Switch to building boringssl with the library.mk system instead of as a
generic lk module.

Test: build.py qemu-generic-arm64-test-debug
Change-Id: I04b91b6c18480bdc9018be18aebaf9b40bdefedf

*external/freetype updated to 7e685b031db8143670fe181e376f57e85ba7e0de
Merge "[trusty] Add teeui to library build system"
*external/libcppbor updated to 10ca06966bbd9c461c12d814b55b1d8f2902a174
Add libcppcbor to library build system

Switch to building libcppcbor with the library.mk system instead of as a
generic lk module.

Test: build.py qemu-generic-arm64-test-debug
Change-Id: I04b91b6c18480bdc9018be18aebaf9b40bdefedf

*system/gatekeeper updated to e743222f7f0757da0b5d41145d73534a6210e0db
Add gatekeeper to userspace library build system

Switch to building system/gatekeeper with the library.mk system instead
of as a generic lk module.

Test: build.py qemu-generic-arm64-test-debug
Change-Id: I2768e500cf28940ca2225b057e594593c06b251b