Apply CleanPath() before comparing paths
Paths are sometimes prefixed with "./". CleanPath() is moved from
FakeIoDelegate to IoDelegate so that Parser uses it before comparing
paths.

Ideally, we should use std::filesystem::path::lexically_normal(), but
it's not available on windows platform which we support.

Removing "./" is good enough for most cases. To be specific, this
happens when we pass "-I.".

Bug: n/a
Test: m (aidl_unittests)
Change-Id: Ied6eaf1b264102d997306ac0cc2c32934056b27c