abitidy: sort all abi-corpus-group abi-corpus abi-instr together
The --sort option combines all abi-instr within an abi-corpus into a
single abi-instr containing the namespaces, types and declarations in
sorted order.

In the case of an abi-corpus-group, each abi-corpus was treated
separately. This commit makes --sort merge all the abi-instr across
all the corpora in a corpus group into a single replacement abi-instr
in the first corpus of the group.

	* tools/abitidy.cc (sort_namespaces_types_and_declarations):
	Change handling of abi-corpus-group so that all abi-instr are
	collected together and the replacement abi-instr is added to
	the first corpus of the group.

Bug: 181269148
Change-Id: I2811efae6f044fbc5311b6a0d854364dfedb1142
Signed-off-by: Giuliano Procida <gprocida@google.com>