UI: Add initial puppeteer-based integrationtests
This CL introduces support for UI integrationtests.
Those tests are based on puppeteer (Chrome headless)
and work by injecting user actions into the real UI,
capturing screenshots and diffing them against
reference ones.

Each test() unit in ui_integrationtests.ts automatically
captures and diffs the screenshot at the end of the run.
In case of failures, diffs are uploaded to the CI GCS
bucket (gs://perfetto-ci-artifacts/$key/).

I got to a state where tests can be run both on Linux
and Mac successfully. There are few rendering differences
between the two, due to the fact that font rendering
seems platform specific. But with enough tweaks the
difference can be contained to sub-pixel antialiasing
and can be dealt with a threshold in the diff algorithm.

Bug: 190075400
Change-Id: I5fb35c41ba5f95a4761465b9d0128edd92414510