Autoupdate of 4 submodules
*art updated to 17374091085affd929d03664bee5a827e76b1b21
Be more prescise in the profile analysis

Distinguish between:
 - compile
 - don't compile because we don't have enough delta (standard case)
 - don't compile because all profiles are empty

This will help us be more precise in the compilation strategy and
avoid re-compiling verified dex files when the profile is empty.

Also, this CL fixes dexoptanalyzer tests.
We needed to generate odex files instead of oat files in the
dalvik-cache when testing the functionality. That's because during
tests, the parent directory of the apk is always writable and
OatFileAssistant will believe that it needs to select the odex
files despite having an oat file.

The fix caught issues in the downgrade test, which were also
addressed.

Test: gtest
Bug: 188655918
Change-Id: Id8370541f73465b32dc91aeacf2ba4dc2656c290

*cts updated to 22d2e5fbda56041b5794ae92b12b68963f26c234
Merge "Fix the cts compilation tests"
*frameworks/base updated to 4ebbeba91beb71ecb2c6c248082f8a048d9da9ca
Merge changes from topics "delete_odex", "verify-prof"

* changes:
  Return the freed bytes from deleteOdex API
  Don't recompile apks with speed-profile is the profile is empty

*frameworks/native updated to d8ee0ace65b784dbdb898eaf19971e17718a6110
Merge changes from topics "delete_odex", "verify-prof"

* changes:
  Return the freed bytes from deleteOdex API
  Extend profile analysis with proper return codes