Further improve 2040-huge-native-alloc reliability
Add trace variant to knownfailures.

Add a way to query the current GC counter. Use that to retry in the
event of an unexpectedly early GC, and to fail if we don't see
a newly completed GC.

Does not yet address "excessive blocking" issue we've seen
occasionally. The addition of a less timing-sensitive success
criterion does give us more options.

Test: Check retry path with temporary code change; Treehugger.
Bug: 190148586
Change-Id: Ide2f6f721a86296042fcc66705cc110335f3df9b