Autoupdate of 2 submodules
*device/google/bonito updated to e0a7d150c2716d5fa8048987e1eb73b37ac730c7
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: Ia14ead102f1751477f99af284e7da192351fe2c1

*device/google/crosshatch updated to 550f00a59d3ca50ba00ea550e6499585a2cbef19
Merge "Disable dexpreopt and <uses-library> check for APKs."