Autoupdate of 6 submodules
*device/google/bonito updated to e0a7d150c2716d5fa8048987e1eb73b37ac730c7
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: Ia14ead102f1751477f99af284e7da192351fe2c1

*device/google/bramble updated to 21aa025bf7aef084f6bad1833d3a0591334d58de
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: I81a4a58b601b3fb496ae7b78eb6abaff20dbf451

*device/google/coral updated to 03da6a8a3e293cab8529a6d711f0a8139dc02bc6
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: Idfa03705f7cbafc70036d323c90450dae99e39f1

*device/google/crosshatch updated to 550f00a59d3ca50ba00ea550e6499585a2cbef19
Merge "Disable dexpreopt and <uses-library> check for APKs."
*device/google/redfin updated to 64937a76ff0cd1af010f23d501a3b6fd4f3dce4d
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: I2f69ba7b8d11e22496cc27a67733515a5d0f1395

*device/google/sunfish updated to 2424cb9b5b48fbb7d1ea8467516b6463fbf2759e
Disable dexpreopt and <uses-library> check for APKs.

This has to be done because the APKs depend on libraries for which there
are no modules in the build system, so it's impossible to add a build
dependency on them. As a consequence, it is not possible to construct
class loader context for dexpreopt.

Bug: 132357300
Test: treehugger
Change-Id: I6891ff2cc7857263ec1fc0f17585b679ef019f5e