Make sure all logging is using the native log code.
As the unwinder is being adopted by other systems, it's necessary to
isolate the Android dependent parts. The code was already written to
isolate the log code out, but a few native Android log calls slipped
into the build, so this CL fixes all of that. It also adds a new build
target to make sure that no new Android log references get in to the
unwinder.

Test: Builds and unit tests pass.
Change-Id: I6c28568c35c867bae4725086b775beeb3d57b0cb