Increase reliability of tradefed media preparer
tradefed media preparer assumes that it the cached directory is still
present, all the contents are present. This is untrue in a linux or mac
environment where /tmp is regularly pruned for old, unaccessed files.

This generates a table of contents (TOC) when the media is first
unzipped. This TOC is then used to verify that all files are still
present when the preparer is invoked; if not, it ignores this cached
copy and re-downloads & unpacks the specified zip file.

Added boolean option "simple-caching-semantics" to support running with the old
caching semantics. Defaults 'false' to use the new TOC-based scheme.

Bug: 176119883
Test: cts-tradefed with various combinations of corrupting host cache
Change-Id: I49f27397980acdf367ea312bdfb1abe83684a347